Group code into files according to relevance
